refactor: storage drivers
Previously, the Storage Drivers were called Api Client because the first
of them was the Rest Api Client which fetched data from and stored  data
to the remote server. It makes more  sense  to  label  them  as  Storage
Drivers because they primary purpose is to persist data, not  necessarly
via an API, such as the Local Storage Driver.